Another pass at getting things working on x64.  This change incorporates most of Sidnei's work on the AMD64 branch, and updates most of the other win32 and win32com modules that haven't already had 64bit love from Roger (thanks guys!).  Note this is not complete - among the outstanding issues are fixing 's#sf' format strings (but most of the tests *do* pass on x64, and the ones which don't fail for 'vista environment' reasons rather than x64 reasons)

+++ com/win32com/src/extensions/PyFUNCDESC.cpp
@@ -72,7 +72,7 @@ BOOL PyObject_AsFUNCDESC(PyObject *ob, FUNCDESC **ppfd)
 		}
 		fd->cScodes = (short)PySequence_Length(pyfd->scodeArray);
 		fd->lprgscode = (SCODE *)AllocMore(fd, sizeof(SCODE)*fd->cScodes);
-		for (int i=0;i<fd->cScodes;i++) {
+		for (Py_ssize_t i=0;i<fd->cScodes;i++) {
 			PyObject *sub=PySequence_GetItem(ob, i);
 			if (sub==NULL) goto done;
 			BOOL ok = PyInt_Check(sub);
@@ -219,14 +219,14 @@ int PyFUNCDESC::setattr(PyObject *self, char *name, PyObject *v)
 }
 
 // Sequence stuff to provide compatibility with tuples.
-/* static */ int PyFUNCDESC::getlength(PyObject *self)
+/* static */ Py_ssize_t PyFUNCDESC::getlength(PyObject *self)
 {
 	// NEVER CHANGE THIS - you will break all the old
 	// code written when these object were tuples!
 	return 10;
 }
 
-/* static */ PyObject *PyFUNCDESC::getitem(PyObject *self, int index)
+/* static */ PyObject *PyFUNCDESC::getitem(PyObject *self, Py_ssize_t index)
 {
 	PyFUNCDESC *p = (PyFUNCDESC *)self;
 	PyObject *rc;
