+++ win32/src/PyDEVMODE.cpp
@@ -277,7 +277,7 @@ PyObject *PyDEVMODE::getattro(PyObject *self, PyObject *obname)
 int PyDEVMODE::setattro(PyObject *self, PyObject *obname, PyObject *obvalue)
 {
 	char *name, *value;
-	int valuelen;
+	Py_ssize_t valuelen;
 	name=PyString_AsString(obname);
 	if (name==NULL)
 		return -1;
